seperate header from navigation links
Importantly added a '!flex' to the carousel. This makes prevents swiper from automatically making this a 'block' element.

And use the semantic 'nav'
From the spec (and from the SO issue that I found)
The nav element represents a section of a page that links to other pages or to parts within the page: a section with navigation links.